# target/classes/api.yaml
swagger: "2.0"
info:
  description: ""
  version: "1.0.0"
  title: "Swagger Payment"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
basePath: "/"
schemes:
  - "https"
  - "http"
paths:
  /payments:
    post:
      tags:
        - "insert"
      summary: "Add a new payment"
      description: ""
      operationId: "addPayment"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Payment object that needs to be added"
          required: true
          schema:
            $ref: "#/definitions/Payment"
      responses:
        "405":
          description: "Invalid input"
      security:
        - payment_auth:
            - "write:payment"
            - "read:payment"
    get:
      tags:
        - "get"
      summary: "Consult payment"
      description: ""
      operationId: "payment"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Payment object that needs to be added"
          required: true
          schema:
            $ref: "#/definitions/Payment"
      responses:
        "405":
          description: "Invalid input"
      security:
        - payment_auth:
            - "write:payment"
            - "read:payment"
    delete:
      tags:
        - "delete"
      summary: "Delete payment"
      description: ""
      operationId: "delete"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Payment object that needs to be added"
          required: true
          schema:
            $ref: "#/definitions/Payment"
      responses:
        "405":
          description: "Invalid input"
      security:
        - payment_auth:
            - "write:payment"
            - "read:payment"
    put:
      tags:
        - "update"
      summary: "update payment"
      description: ""
      operationId: "updatePayment"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Payment object that needs to be added"
          required: true
          schema:
            $ref: "#/definitions/Payment"
      responses:
        "405":
          description: "Invalid input"
      security:
        - payment_auth:
            - "write:payment"
            - "read:payment"


securityDefinitions:
  payment_auth:
    type: "oauth2"
    authorizationUrl: "http://payment.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:payment: "create a new payment"
      read:payment: "read your payments"
definitions:
  Payment:
    type: "object"
    required:
      - "inclusionDate"
      - "date"
      - "value"
      - "description"
    properties:
      id:
        type: "integer"
        format: "int64"
      status:
        type: "string"
        description: "payment status"
        enum:
          - "AGENDADO"
          - "EFETUADO"
          - "PENDENTE"
          - "CANCELADO"
      inclusionDate:
        type: "string"
        format: "date"
      date:
        type: "string"
        format: "date"
      value:
        type: "number"
      description:
        type: "string"
      recurrence:
        $ref: "#/definitions/Recurrence"
      destination:
        $ref: "#/definitions/Destination"
  Recurrence:
    type: "object"
    properties:
      finalDate:
        type: "string"
        format: "date"
      frequency:
        description: "payment status"
        enum:
          - "SEMANAL"
          - "MENSAL"
          - "TRIMESTRAL"
          - "SEMESTRAL"
  Destination:
    type: "object"
    required:
      - "pixKey"
    properties:
      pixKey:
        type: "string"
      keyType:
        description: "payment status"
        enum:
          - "CPF"
          - "E-MAIL"
          - "TELEFONE"
          - "ALEATORIA"

externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"